refactor(cordova-validator): saner names
