+++ lib/commands/build.js
@@ -3,8 +3,8 @@
 var CdvBuildTask    = require('../tasks/cordova-build');
 var HookTask        = require('../tasks/run-hook');
 var BuildTask       = require('../tasks/ember-build');
-var ValidatePlatformTask = require('../tasks/platform-exists');
 
+var ValidatePlatformTask    = require('../tasks/validate-platform');
 var validateLocationType    = require('../utils/validate-location-type');
 var getCordovaConfig        = require('../utils/get-cordova-config');
 var validateAllowNavigation = require('../utils/validate-allow-navigation');
