+++ lib/commands/serve.js
@@ -3,13 +3,14 @@
 var CdvBuildTask    = require('../tasks/cordova-build');
 var HookTask        = require('../tasks/run-hook');
 var ServeTask       = require('../tasks/ember-build-serve');
-var ValidatePluginTask = require('../tasks/plugin-exists');
 
 var Promise         = require('ember-cli/lib/ext/promise');
 var PortFinder      = require('portfinder');
 var chalk           = require('chalk');
 var _merge          = require('lodash').merge;
 
+var ValidatePlatformTask    = require('../tasks/validate-platform');
+var ValidatePluginTask      = require('../tasks/validate-plugin');
 var validateLocationType    = require('../utils/validate-location-type');
 var getCordovaConfig        = require('../utils/get-cordova-config');
 var validateAllowNavigation = require('../utils/validate-allow-navigation');
@@ -51,7 +52,7 @@ module.exports = {
   /* eslint-enable max-len */
 
   run: function(options) {
-    var hook, serve, cordovaBuild, validateCordovaConfig, pluginExists;
+    var hook, serve, cordovaBuild, validateCordovaConfig, pluginExists, platformExists;
 
     validateLocationType(this.project.config());
 
@@ -74,6 +75,12 @@ module.exports = {
       pluginName: 'cordova-plugin-whitelist'
     });
 
+    platformExists = new ValidatePlatformTask({
+      project: this.project,
+      ui: this.ui,
+      platform: platform
+    });
+
     hook = new HookTask({
       project: this.project,
       ui: this.ui
@@ -91,6 +98,7 @@ module.exports = {
     });
 
     return validateCordovaConfig
+      .then(platformExists.prepare())
       .then(pluginExists.prepare())
       .then(hook.prepare('beforeBuild'))
       .then(this._autoFindLiveReloadPort.bind(this, options))
