+++ lib/tasks/validate-platform.js
