+++ lib/tasks/validate-plugin.js
